better README and description, possible variants for enums, shell completion
