+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod annotations;
-#[cfg(feature = "application")]
+#[cfg(feature = "cli")]
 pub mod cli;
 pub mod errors;
 pub mod groups;
